0.23078 más correcciones en bonos

+++ model/Bono.go
@@ -3,6 +3,6 @@ package model
 // Tbono es la estructura para usuario
 type Tbono struct {
 	ID       int64
-	Precio   float32
+	Precio   float64
 	Sesiones int
 }
